pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bank-fraud-detection"
version = "0.1.0"
description = "Pipeline Big Data et détection de fraude sur transactions bancaires (Kafka, ML, API)."
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "numpy==1.26.4",
    "pandas==2.2.3",
    "pyarrow==18.1.0",
    "joblib==1.4.2",
    "scikit-learn==1.5.2",
    "imbalanced-learn==0.12.4",
    "xgboost==2.1.2",
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.9.2",
    "kafka-python==2.0.2",
]

[project.optional-dependencies]
notebooks = [
    "matplotlib==3.9.2",
    "seaborn==0.13.2",
    "jupyterlab==4.3.1",
    "ipykernel==6.29.5",
]

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["bank_fraud_detection*", "simulateur"]
